% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{convert_stratification_factors}
\alias{convert_stratification_factors}
\title{Convert stratification factors}
\usage{
convert_stratification_factors(trial_data)
}
\arguments{
\item{trial_data}{dataframe of other variables}
}
\description{
Convert stratification factors to be consistent in their levels and names,
and add indicator variables
}
