% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_characteristics.R
\name{create_characteristics_table}
\alias{create_characteristics_table}
\title{Table of participant characteristics}
\usage{
create_characteristics_table(trial_mod)
}
\arguments{
\item{trial_mod}{dataframe of all variables collected}
}
\description{
Table of participant characteristics
}
