% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{create_codebook}
\alias{create_codebook}
\title{Codebook of variables}
\usage{
create_codebook(trial_datatable)
}
\arguments{
\item{trial_datatable}{Dataframe of all variables collected}
}
\description{
Codebook of variables
}
