% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consort.R
\name{create_consort_pdf}
\alias{create_consort_pdf}
\title{CONSORT diagram}
\usage{
create_consort_pdf(screen, trial)
}
\arguments{
\item{screen}{dataframe of screening inclusions and exclusions}

\item{trial}{dataframe of participants}
}
\description{
CONSORT diagram
}
