% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{create_consort_png}
\alias{create_consort_png}
\title{CONSORT diagram}
\usage{
create_consort_png(screen, trial)
}
\arguments{
\item{screen}{dataframe of screening inclusions and exclusions}

\item{trial}{dataframe of participants}
}
\description{
CONSORT diagram
}
