% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen_use.R
\name{create_fio2}
\alias{create_fio2}
\title{Dataframe of fio2 settings}
\usage{
create_fio2(oxygen)
}
\description{
Dataframe of all columns needed from oxygen dataframe (participant id, all
oxygen fio2 values used, all time intervals between changes in oxygen fio2
value, and assigned randomization of each high flow nasal oxygen
participant)
}
