% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen_use.R
\name{create_fio2_table}
\alias{create_fio2_table}
\title{fio2 table in wide format}
\usage{
create_fio2_table(fio2)
}
\description{
Wide dataframe of all oxygen fio2 values at every minute for each high flow
}
