% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen_use.R
\name{create_fio2_table_long}
\alias{create_fio2_table_long}
\title{fio2 table in long format}
\usage{
create_fio2_table_long(fio2, fio2_table)
}
\description{
Merge all oxygen fio2 values into one column named "fio2"
}
