% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_med_summary.R
\name{create_med_summary}
\alias{create_med_summary}
\title{Medication summary}
\usage{
create_med_summary(trial_mod)
}
\arguments{
\item{trial_mod}{dataframe of all variables collected}
}
\description{
Medication summary
}
