% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_outcome_plots.R
\name{create_ordinal_outcome_plot}
\alias{create_ordinal_outcome_plot}
\title{Difference between groups for ordinal outcomes}
\usage{
create_ordinal_outcome_plot(trial_mod, factor)
}
\arguments{
\item{trial_mod}{dataframe}

\item{factor}{ordinal outcome to plot}
}
\description{
Difference between groups for ordinal outcomes
}
