% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen_use.R
\name{create_oxygen}
\alias{create_oxygen}
\title{Create a table containing variables related to measuring oxygen device settings}
\usage{
create_oxygen(trial_mod)
}
\description{
Dataframe of all columns needed from trial dataframe (participant id, all
oxygen flow rates used, all oxygen fio2 values used, and assigned
randomization of each participant), with all time intervals between
changes in oxygen flow rate/oxygen fio2 value
}
