% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen_use.R
\name{create_oxygen_fio2_plot}
\alias{create_oxygen_fio2_plot}
\title{fio2 table in long format}
\usage{
create_oxygen_fio2_plot(fio2_table_long)
}
\description{
fio2 values used in sequence during procedure for participants randomized to HFNO
}
