% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen_use.R
\name{create_oxygen_flow_plot}
\alias{create_oxygen_flow_plot}
\title{Oxygen flow rate sequence plot}
\usage{
create_oxygen_flow_plot(oxygen_table_long)
}
\description{
Plot of all oxygen flow rate values used in sequence during procedures
}
