% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen_use.R
\name{create_oxygen_proportion_plot}
\alias{create_oxygen_proportion_plot}
\title{Oxygen flow rate stacked plot}
\usage{
create_oxygen_proportion_plot(oxygen_table_long)
}
\description{
Plot of proportion of time spent during procedure using different oxygen
flow rates
}
