% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen_use.R
\name{create_oxygen_table}
\alias{create_oxygen_table}
\title{Oxygen flow rate table in wide format}
\usage{
create_oxygen_table(oxygen)
}
\description{
Dataframe of all oxygen flow rate values at every minute, with assigned
randomization of each participant
}
