% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen_use.R
\name{create_oxygen_table_long}
\alias{create_oxygen_table_long}
\title{Oxygen flow rate table in long format}
\usage{
create_oxygen_table_long(oxygen, oxygen_table)
}
\description{
Merge all oxygen flow rate values into one column named "flowrate"
}
