% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_spo2.R
\name{create_spo2_plot}
\alias{create_spo2_plot}
\title{Plot to explore amount of missing spo2 data per participant}
\usage{
create_spo2_plot(co2_long)
}
\arguments{
\item{tcco2_data}{dataframe of co2, spo2, pulse rate}

\item{trial_mod}{dataframe of all other variables collected}
}
\description{
Plot to explore amount of missing spo2 data per participant
}
