% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_spo2.R
\name{create_tcco2_data}
\alias{create_tcco2_data}
\title{Load data from tcco2 monitor}
\usage{
create_tcco2_data(TCCO2)
}
\arguments{
\item{data}{Filepath to directory of csv files}
}
\description{
Load data from tcco2 monitor
}
