% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_timesusedhfno_plot.R
\name{create_timesusedhfno_plot}
\alias{create_timesusedhfno_plot}
\title{Plot of number of times Anesthesia Assistants used HFNO}
\usage{
create_timesusedhfno_plot(trial_mod)
}
\arguments{
\item{trial_mod}{dataframe of all variables collected}
}
\description{
Plot of number of times Anesthesia Assistants used HFNO
}
