% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_datatable.R
\name{create_trial_datatable}
\alias{create_trial_datatable}
\title{Selection of all variables used in the analysis}
\usage{
create_trial_datatable(trial_mod, data_labelled)
}
\arguments{
\item{trial_mod}{Dataframe of all variables collected}

\item{data_labelled}{Dataframe with labels of variables}
}
\description{
Selection of all variables used in the analysis
}
