% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_models.R
\name{extract_CI}
\alias{extract_CI}
\title{Credible intervals for treatment effect}
\usage{
extract_CI(
  model,
  unit = NULL,
  pars = "randomization_factorHighFlownasaloxygen"
)
}
\description{
Credible intervals for treatment effect
}
