% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_CI_brms.R
\name{extract_CI_brms}
\alias{extract_CI_brms}
\title{Credible intervals for treatment effect}
\usage{
extract_CI_brms(
  model_obj,
  unit = NULL,
  pars = "randomization_factorHighFlownasaloxygen",
  effect_type,
  logarithm = FALSE
)
}
\description{
Credible intervals for treatment effect
}
