% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_inla_CI.R
\name{extract_inla_CI}
\alias{extract_inla_CI}
\title{Credible intervals for FANOVA}
\usage{
extract_inla_CI(
  model_obj,
  unit = NULL,
  pars = "randomization_factorHighFlownasaloxygen"
)
}
\description{
Credible intervals for FANOVA
}
