% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_effect_modification}
\alias{fit_effect_modification}
\title{Primary outcome with effect modification for OSA and CRT}
\usage{
fit_effect_modification(data_primary, response)
}
\arguments{
\item{data_primary}{dataframe}

\item{response}{peak or mean}
}
\description{
Used for "subgroup analysis"
}
