% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_primary}
\alias{fit_primary}
\title{Fit model for tcco2 outcomes}
\usage{
fit_primary(data_primary, response)
}
\arguments{
\item{data_primary}{dataframe}

\item{response}{peak or mean}
}
\description{
Fit model for tcco2 outcomes
}
