% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_propodds}
\alias{fit_propodds}
\title{Fit proportional odds models}
\usage{
fit_propodds(data_propodds, response)
}
\arguments{
\item{data_propodds}{dataframe}

\item{response}{outcome variable}
}
\description{
Fit proportional odds models
}
