% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_troops}
\alias{fit_troops}
\title{Fit model for TROOPS}
\usage{
fit_troops(data_troops)
}
\arguments{
\item{data_troops}{dataframe}
}
\description{
Tracking and reporting outcomes of procedural sedation
}
