% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_brms_summary.R
\name{make_brms_summary}
\alias{make_brms_summary}
\title{Summary of brms models}
\usage{
make_brms_summary(model_obj, unit, effect_type, response, response_title)
}
\description{
Summary of brms models
}
