% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{make_co2_long}
\alias{make_co2_long}
\title{Combine co2 and trial data into a long-format dataframe}
\usage{
make_co2_long(co2_data, trial_data)
}
\arguments{
\item{co2_data}{dataframe of data from tcco2 monitor}

\item{trial_data}{dataframe of other variables}
}
\description{
Combine co2 and trial data into a long-format dataframe
}
