% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flextable.R
\name{make_flextable_outcomes_formatted}
\alias{make_flextable_outcomes_formatted}
\title{Format outcomes for table}
\usage{
make_flextable_outcomes_formatted(model_list, data_list)
}
\description{
Format outcomes for table
}
