% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_plot_co2_mean}
\alias{make_plot_co2_mean}
\title{Plot mean co2}
\usage{
make_plot_co2_mean(data_primary, model)
}
\arguments{
\item{data_primary}{dataframe}

\item{model}{model object}
}
\description{
Plots mean co2 against baseline
}
