% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_plot_co2_peak}
\alias{make_plot_co2_peak}
\title{Plot peak co2}
\usage{
make_plot_co2_peak(data_primary, model)
}
\arguments{
\item{data_primary}{dataframe}

\item{model}{model object}
}
\description{
Plots peak co2 against baseline
}
