% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_component.R
\name{make_plot_component}
\alias{make_plot_component}
\title{FANOVA effect plot}
\usage{
make_plot_component(
  data_fanova,
  model,
  reso,
  component_str,
  title,
  ylim = c(-40, 40)
)
}
\description{
FANOVA effect plot
}
