% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spo2_long.R
\name{make_spo2_long}
\alias{make_spo2_long}
\title{Combine spo2 and trial data into a long-format dataframe}
\usage{
make_spo2_long(spo2_data, trial_data)
}
\arguments{
\item{spo2_data}{dataframe of data from anesthesia monitor}

\item{trial_data}{dataframe of other variables}
}
\description{
Combine spo2 and trial data into a long-format dataframe
}
