% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary_descr.R
\name{make_summary_descr}
\alias{make_summary_descr}
\title{Summary description}
\usage{
make_summary_descr(
  data_obj,
  unit,
  response,
  response_title,
  factor_levels = NULL
)
}
\description{
Summary description
}
