% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_outcomes_formatted.R
\name{make_table_outcomes_formatted}
\alias{make_table_outcomes_formatted}
\title{Format outcomes for table}
\usage{
make_table_outcomes_formatted(model_list, data_list)
}
\description{
Format outcomes for table
}
