% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prob_summary.R
\name{posterior_prob_summary}
\alias{posterior_prob_summary}
\title{Summaries from posterior probabilities}
\usage{
posterior_prob_summary(
  model_obj,
  unit = NULL,
  pars = "randomization_factorHighFlownasaloxygen",
  threshold
)
}
\description{
Summaries from posterior probabilities
}
