% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_statements.R
\name{prob_threshold}
\alias{prob_threshold}
\title{Probability above/in/below threshold}
\usage{
prob_threshold(model_obj, thresh = 0, pars = "randomization_factor")
}
\arguments{
\item{model_obj}{thresh pars}

\item{response}{peak or mean}
}
\description{
Probability above/in/below threshold
}
