% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assist.R
\name{process_assist}
\alias{process_assist}
\title{Process data for anesthesia assistant ratings analysis}
\usage{
process_assist(trial_data)
}
\arguments{
\item{trial_data}{dataframe of other variables}
}
\description{
Ratings for ease of use of oxygen device
}
