% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_comfort}
\alias{process_comfort}
\title{Process data for patient comfort analysis}
\usage{
process_comfort(trial_data)
}
\arguments{
\item{trial_data}{dataframe of other variables}
}
\description{
Comfort with oxygen device
}
