% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_primary.R
\name{process_primary}
\alias{process_primary}
\title{Process data for Primary analysis}
\usage{
process_primary(co2_long)
}
\arguments{
\item{co2_data}{dataframe of data from tcco2 monitor}

\item{trial_data}{dataframe of other variables}
}
\description{
Baseline is taken as first non-missing co2 value. Peak is taken as largest co2 value
}
