% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_spo2.R
\name{process_spo2}
\alias{process_spo2}
\title{Process data for SPO2 analysis}
\usage{
process_spo2(spo2_trial)
}
\arguments{
\item{spo2_trial}{long dataframe of data from anesthesia monitor
and trial data}
}
\description{
Outcome to assess oxygenation
}
