% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_troops.R
\name{process_troops}
\alias{process_troops}
\title{Process data for TROOPS analysis}
\usage{
process_troops(trial_data)
}
\arguments{
\item{trial_data}{dataframe of other variables}
}
\description{
Tracking and reporting outcomes of procedural sedation
}
