% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_date.R
\name{time_to_date}
\alias{time_to_date}
\title{Convert Time to a datetime object}
\usage{
time_to_date(data)
}
\arguments{
\item{data}{Used for co2_data}
}
\description{
Date format is mixed, so we have to treat each format separately
}
