% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_fanova_data.R
\name{make_plot_fanova_data}
\alias{make_plot_fanova_data}
\title{Plot fanova data}
\usage{
make_plot_fanova_data(data_fanova, reso)
}
\arguments{
\item{data_fanova}{dataframe}

\item{reso}{Number of seconds resolution for functional data}
}
\description{
Plot fanova data
}
