% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_fanova}
\alias{process_fanova}
\title{Process data for FANOVA analysis}
\usage{
process_fanova(co2_data, trial_data, reso = 30)
}
\arguments{
\item{co2_data}{dataframe of data from tcco2 monitor}

\item{trial_data}{dataframe of other variables}

\item{reso}{Number of seconds resolution for functional data}
}
\description{
Process data for FANOVA analysis
}
