% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_spo2.R
\name{process_spo2}
\alias{process_spo2}
\title{Process data for SPO2 analysis}
\usage{
process_spo2(co2_long)
}
\arguments{
\item{trial_data}{dataframe of other variables}

\item{co2_data}{dataframe of data from tcco2 monitor}
}
\description{
Outcome to assess oxygenation
}
