% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{process_spo2}
\alias{process_spo2}
\title{Process data for SPO2 analysis}
\usage{
process_spo2(co2_data, trial_data)
}
\arguments{
\item{co2_data}{dataframe of data from tcco2 monitor}

\item{trial_data}{dataframe of other variables}
}
\description{
Outcome to assess oxygenation
}
